Remove the TemplateStatisticStrategy for enhanced performance and simplified design
Eliminate TemplateMethod by moving its only shared step (prepareDataset) to CourseStatisticsService. This change allows for the sharing of a single DescriptiveStatistic object (the prepared dataset - that does the stats calculations) among all StatisticStrategy instances and leads to improved performance and simplified design.
