+++ src/main/java/com/myy803/course_mgt_app/service/statistics/MeanStatisticStrategy.java
@@ -4,14 +4,14 @@ import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
 import org.springframework.stereotype.Component;
 
 @Component
-public class MeanStatisticStrategy extends TemplateStatisticStrategy{
+public class MeanStatisticStrategy extends StatisticStrategy{
 	
 	public MeanStatisticStrategy() {
 		super("Mean");	// name of the statistic
 	}
 
 	@Override
-	public double doActualCalculation(DescriptiveStatistics ds) {
+	public double calculateStatistic(DescriptiveStatistics ds) {
 		return ds.getMean();
 	}
