+++ src/main/java/com/myy803/course_mgt_app/service/statistics/MinStatisticStrategy.java
@@ -4,14 +4,14 @@ import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
 import org.springframework.stereotype.Component;
 
 @Component
-public class MinStatisticStrategy extends TemplateStatisticStrategy{
+public class MinStatisticStrategy extends StatisticStrategy{
 	
 	public MinStatisticStrategy() {
 		super("Min");
 	}
 
 	@Override
-	public double doActualCalculation(DescriptiveStatistics ds) {
+	public double calculateStatistic(DescriptiveStatistics ds) {
 		return ds.getMin();
 	}
