+++ src/main/java/com/myy803/course_mgt_app/service/statistics/PercentileStatisticStrategy.java
@@ -4,14 +4,14 @@ import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
 import org.springframework.stereotype.Component;
 
 @Component
-public class PercentileStatisticStrategy extends TemplateStatisticStrategy{
+public class PercentileStatisticStrategy extends StatisticStrategy{
 
 	public PercentileStatisticStrategy() {
 		super("Percentile");
 	}
 
 	@Override
-	public double doActualCalculation(DescriptiveStatistics ds) {
+	public double calculateStatistic(DescriptiveStatistics ds) {
 		return ds.getPercentile(50);
 	}
