+++ src/main/java/com/myy803/course_mgt_app/service/statistics/StandardDeviationStatisticStrategy.java
@@ -4,14 +4,14 @@ import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
 import org.springframework.stereotype.Component;
 
 @Component
-public class StandardDeviationStatisticStrategy extends TemplateStatisticStrategy {
+public class StandardDeviationStatisticStrategy extends StatisticStrategy {
 
 	public StandardDeviationStatisticStrategy() {
 		super("StandardDeviation");	// name of the statistic
 	}
 
 	@Override
-	public double doActualCalculation(DescriptiveStatistics ds) {
+	public double calculateStatistic(DescriptiveStatistics ds) {
 		return ds.getStandardDeviation();
 	}
