+++ src/main/java/com/myy803/course_mgt_app/service/statistics/StatisticStrategy.java
@@ -1,9 +1,22 @@
 package com.myy803.course_mgt_app.service.statistics;
 
-import java.util.List;
+import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
 
-public interface StatisticStrategy {
+public abstract class StatisticStrategy {
+	
+	protected String statisticName; 
+	
+	public StatisticStrategy() {}
+	
+	public StatisticStrategy(String statisticName) {
+		super();
+		this.statisticName = statisticName;
+	}
 
-	Double calculateStatistic(List<Double> grades);
+	public String getStatisticName() {
+		return statisticName;
+	}
+	
+	abstract public double calculateStatistic(DescriptiveStatistics ds);
 
 }
